Rmpfr-package {Rmpfr}	R Documentation

R MPFR - Multiple Precision Floating-Point Reliable

Description

Rmpfr provides S4 classes and methods for arithmetic including transcendental ("special") functions for arbitrary precision floating point numbers, here often called “mpfr - numbers”. To this end, it interfaces to the LGPL'ed MPFR (Multiple Precision Floating-Point Reliable) Library which itself is based on the GMP (GNU Multiple Precision) Library.

Details

Package: 	 Rmpfr

Title: 	 Interface R to MPFR - Multiple Precision Floating-Point Reliable

Version: 	 1.1-2

VersionNote: 	 Last CRAN: 1.1-1 on 2025-07-18; 1.1-0 on 2025-05-08; 1.0-0 on 2024-11-15

Date: 	 2025-10-21

Type: 	 Package

Authors@R: 	 c(person("Martin","Maechler", role = c("aut","cre"), email = "maechler@stat.math.ethz.ch", comment = c(ORCID="0000-0002-8685-9910")) , person(c("Richard", "M."), "Heiberger", role = "ctb", email="rmh@temple.edu", comment = "formatHex(), *Bin, *Dec") , person(c("John", "C."), "Nash", role = "ctb", email="nashjc@uottawa.ca", comment = "hjkMpfr(), origin of unirootR()") , person(c("Hans", "W."), "Borchers", role = "ctb", email="hwborchers@googlemail.com", comment = "optimizeR(*, \"GoldenRatio\"); origin of hjkMpfr()") , person("Mikael", "Jagan", role = "ctb", comment = c("safer convert.c; configure.ac etc", ORCID = "0000-0002-3542-2938")) )

Description: 	 Arithmetic (via S4 classes and methods) for arbitrary precision floating point numbers, including transcendental ("special") functions. To this end, the package interfaces to the 'LGPL' licensed 'MPFR' (Multiple Precision Floating-Point Reliable) Library which itself is based on the 'GMP' (GNU Multiple Precision) Library.

SystemRequirements: 	 gmp (>= 4.2.3), mpfr (>= 3.2.0), pdfcrop (part of TexLive) is required to rebuild the vignettes.

SystemRequirementsNote: 	 'MPFR' (MP Floating-Point Reliable Library, https://www.mpfr.org/) and 'GMP' (GNU Multiple Precision library, https://gmplib.org/), see >> README.md

Depends: 	 gmp (>= 0.6-1), R (>= 3.6.0)

Imports: 	 stats, utils, methods

Suggests: 	 DPQmpfr, MASS, Bessel, polynom, sfsmisc (>= 1.1-14)

SuggestsNote: 	 MASS, polynom, sfsmisc: only for vignette;

Enhances: 	 dfoptim, pracma, DPQ

EnhancesNote: 	 mentioned in Rd xrefs | used in example

URL: 	 https://rmpfr.r-forge.r-project.org/

BugReports: 	 https://r-forge.r-project.org/tracker/?group_id=386

License: 	 GPL (>= 2)

Encoding: 	 UTF-8

Author: 	 Martin Maechler [aut, cre] (ORCID: <https://orcid.org/0000-0002-8685-9910>), Richard M. Heiberger [ctb] (formatHex(), *Bin, *Dec), John C. Nash [ctb] (hjkMpfr(), origin of unirootR()), Hans W. Borchers [ctb] (optimizeR(*, "GoldenRatio"); origin of hjkMpfr()), Mikael Jagan [ctb] (safer convert.c; configure.ac etc, ORCID: <https://orcid.org/0000-0002-3542-2938>)

Maintainer: 	 Martin Maechler <maechler@stat.math.ethz.ch>

Index of help topics:

.bigq2mpfr              Conversion Utilities gmp <-> Rmpfr
Bernoulli               Bernoulli Numbers in Arbitrary Precision
Bessel_mpfr             Bessel functions of Integer Order in multiple
                        precisions
Mnumber-class           Class "Mnumber" and "mNumber" of "mpfr" and
                        regular numbers and arrays from them
Rmpfr-package           R MPFR - Multiple Precision Floating-Point
                        Reliable
array_or_vector-class   Auxiliary Class "array_or_vector"
asNumeric-methods       Methods for 'asNumeric(<mpfr>)'
atomicVector-class      Virtual Class "atomicVector" of Atomic Vectors
c.mpfr                  MPFR Number Utilities
cbind                   "mpfr" '...' - Methods for Functions cbind(),
                        rbind()
chooseMpfr              Binomial Coefficients and Pochhammer Symbol aka
                        Rising Factorial
determinant.mpfrMatrix
                        Functions for mpfrMatrix Objects
dnorm                   Distribution Functions with MPFR Arithmetic
factorialMpfr           Factorial 'n!'  in Arbitrary Precision
formatHex               Flexibly Format Numbers in Binary, Hex and
                        Decimal Format
formatMpfr              Formatting and Printing MPFR (multiprecision)
                        Numbers
frexpMpfr               Base-2 Representation and Multiplication of
                        Mpfr Numbers
getPrec                 Rmpfr - Utilities for Precision Setting, etc
hjkMpfr                 Hooke-Jeeves Derivative-Free Minimization R
                        (working for MPFR)
igamma                  Incomplete Gamma Function
integrateR              One-Dimensional Numerical Integration - in pure
                        R
is.whole.mpfr           Whole ("Integer") Numbers
log1mexp                Compute f(a) = log(1 +/- exp(-a)) Numerically
                        Optimally
matmult                 (MPFR) Matrix (Vector) Multiplication
mpfr                    Create "mpfr" Numbers (Objects)
mpfr-class              Class "mpfr" of Multiple Precision Floating
                        Point Numbers
mpfrArray               Construct "mpfrArray" almost as by 'array()'
mpfrMatrix-class        Classes "mpfrMatrix" and "mpfrArray"
num2bigq                BigQ / BigRational Approximation of Numbers
optimizeR               High Precision One-Dimensional Optimization
outer                   Base Functions etc, as an Rmpfr version
pbetaI                  Accurate Incomplete Beta / Beta Probabilities
                        For Integer Shapes
pmax                    Parallel Maxima and Minima
qnormI                  Gaussian / Normal Quantiles 'qnorm()' via
                        Inversion
roundMpfr               Rounding to Binary bits, "mpfr-internally"
sapplyMpfr              Apply a Function over a "mpfr" Vector
seqMpfr                 "mpfr" Sequence Generation
str.mpfr                Compactly Show STRucture of Rmpfr Number Object
sumBinomMpfr            (Alternating) Binomial Sums via Rmpfr
unirootR                One Dimensional Root (Zero) Finding - in pure R
zeta                    Special Mathematical Functions (MPFR)

Further information is available in the following vignettes:

Maechler_useR_2011-abstr 	 useR-2011-abstract (source)

Rmpfr-pkg 	 Arbitrarily Accurate Computation with R Package Rmpfr (source)

log1mexp-note 	 Accurately Computing log(1 - exp(.)) -- Assessed by Rmpfr (source)

The following (help pages) index does not really mention that we provide many methods for mathematical functions, including gamma, digamma, etc, namely, all of R's (S4) Math group (with the only exception of trigamma), see the list in the examples. Additionally also pnorm, the “error function”, and more, see the list in zeta, and further note the first vignette (below).

Partial index:

mpfr 	 Create "mpfr" Numbers (Objects)

mpfrArray 	 Construct "mpfrArray" almost as by array()

mpfr-class 	 Class "mpfr" of Multiple Precision Floating Point Numbers

mpfrMatrix-class 	 Classes "mpfrMatrix" and "mpfrArray"

Bernoulli 	 Bernoulli Numbers in Arbitrary Precision

Bessel_mpfr 	 Bessel functions of Integer Order in multiple precisions

c.mpfr 	 MPFR Number Utilities

cbind 	 "mpfr" ... - Methods for Functions cbind(), rbind()

chooseMpfr 	 Binomial Coefficients and Pochhammer Symbol aka

Rising Factorial

factorialMpfr 	 Factorial 'n!' in Arbitrary Precision

formatMpfr 	 Formatting MPFR (multiprecision) Numbers

getPrec 	 Rmpfr - Utilities for Precision Setting, Printing, etc

roundMpfr 	 Rounding to Binary bits, "mpfr-internally"

seqMpfr 	 "mpfr" Sequence Generation

sumBinomMpfr 	 (Alternating) Binomial Sums via Rmpfr

zeta 	 Special Mathematical Functions (MPFR)

integrateR 	 One-Dimensional Numerical Integration - in pure R

unirootR 	 One Dimensional Root (Zero) Finding - in pure R

optimizeR 	 High Precisione One-Dimensional Optimization

hjkMpfr 	 Hooke-Jeeves Derivative-Free Minimization R (working for MPFR)

Further information is available in the following vignettes:

Rmpfr-pkg 	 Arbitrarily Accurate Computation with R: The 'Rmpfr' package (source, pdf)

log1mexp-note 	 Acccurately Computing log(1 - exp(.)) -- Assessed by Rmpfr (source, pdf)

Author(s)

Martin Maechler

References

MPFR (MP Floating-Point Reliable Library), https://www.mpfr.org/

GMP (GNU Multiple Precision library), https://gmplib.org/

and see the vignettes mentioned above.

See Also

The R package gmp for big integer gmp and rational numbers (bigrational) on which Rmpfr depends.

Examples

## Using  "mpfr" numbers instead of regular numbers...
n1.25 <- mpfr(5, precBits = 256)/4
n1.25

## and then "everything" just works with the desired chosen precision:hig
n1.25 ^ c(1:7, 20, 30) ## fully precise; compare with
print(1.25 ^ 30, digits=19)

exp(n1.25)

## Show all math functions which work with "MPFR" numbers (1 exception: trigamma)
getGroupMembers("Math")

## We provide *many* arithmetic, special function, and other methods:
showMethods(classes = "mpfr")
showMethods(classes = "mpfrArray")

[Package Rmpfr version 1.1-2 Index]