array_or_vector-class {Rmpfr}	R Documentation

Auxiliary Class "array\_or\_vector"

Description

"array_or_vector" is the class union of c("array", "matrix", "vector") and exists for its use in signatures of method definitions.

Details

Using "array_or_vector" instead of just "vector" in a signature makes an important difference: E.g., if we had setMethod(crossprod, c(x="mpfr", y="vector"), function(x,y) CPR(x,y)), a call crossprod(x, matrix(1:6, 2,3)) would extend into a call of CPR(x, as(y, "vector")) such that CPR()'s second argument would simply be a vector instead of the desired 2 x 3 matrix.

Objects from the Class

A virtual Class: No objects may be created from it.

Examples

 showClass("array_or_vector")

[Package Rmpfr version 0.8-2 Index]