igamma {Rmpfr}	R Documentation

Incomplete Gamma Function

Description

For MPFR version >= 3.2.0, the following MPFR library function is provided: mpfr_gamma_inc(a,x), the R interface of which is igamma(a,x), where igamma(a,x) is the “upper” incomplete gamma function

Γ(a,x) :=: Γ(a) - γ(a,x),

where

γ(a,x) := ∫₀ˣ tᵃ⁻¹ e⁻ᵗ dt,

and hence

Γ(a,x) := ∫ (x..∞) tᵃ⁻¹ e⁻ᵗ dt,

and

Γ(a) := γ(a, ∞).

As R's pgamma(x,a) is

\code{pgamma(x, a)} := γ(a,x) / Γ(a),

we get

        igamma(a,x) ==  gamma(a) * pgamma(x, a, lower.tail=FALSE)

Usage

igamma(a, x, rnd.mode = c("N", "D", "U", "Z", "A"))

Arguments

a, x
an object of class mpfr or numeric, where only one of rate and scale should be specified.

rnd.mode
a 1-letter string specifying how rounding should happen at C-level conversion to MPFR, see mpfr.

Value

a numeric vector of “common length”, recyling along a and x.

Warning

The MPFR library documentation on mpfr_gamma_inc() https://www.mpfr.org/mpfr-current/mpfr.html#index-mpfr_005fgamma_005finc contains

Note

the current implementation of mpfr_gamma_inc(rop, op, op2, <rnd>) is slow for large values of rop or op, in which case some internal overflow might also occur.

Author(s)

R interface: Martin Maechler

References

NIST Digital Library of Mathematical Functions, section 8.2. https://dlmf.nist.gov/8.2.i

Wikipedia (2019). Incomplete gamma function; https://en.wikipedia.org/wiki/Incomplete_gamma_function

See Also

R's gamma (function) and pgamma (probability distribution). Rmpfr's own pgamma(), a thin wrapper around igamma().

Examples

## show how close pgamma() is :
x <- c(seq(0,20, by=1/4), 21:50, seq(55, 100, by=5))
if(mpfrVersion() >= "3.2.0") { print(
all.equal(igamma(Const("pi", 80), x),
          pgamma(x, pi, lower.tail=FALSE) * gamma(pi),
          tol=0, formatFUN = function(., ...) format(., digits = 7)) #-> 2.75e-16 (was 3.13e-16)
)
## and ensure *some* closeness:
stopifnot(exprs = {
   all.equal(igamma(Const("pi", 80), x),
             pgamma(x, pi, lower.tail=FALSE) * gamma(pi),
             tol = 1e-15)
})
} # only if MPFR version >= 3.2.0

[Package Rmpfr version 1.1-2 Index]