seqMpfr {Rmpfr}	R Documentation

"mpfr" Sequence Generation

Description

Generate ‘regular’, i.e., arithmetic sequences. This is in lieu of methods for seq (dispatching on all three of from, to, and by.

Usage

seqMpfr(from = 1, to = 1, by = ((to - from)/(length.out - 1)),
        length.out = NULL, along.with = NULL, ...)

Arguments

from, to
the starting and (maximal) end value (numeric or "mpfr") of the sequence.

by
number (numeric or "mpfr"): increment of the sequence.

length.out
desired length of the sequence. A non-negative number, which will be rounded up if fractional.

along.with
take the length from the length of this argument.

...
arguments passed to or from methods.

Details

see seq (default method in package base), whose semantic we want to replicate (almost).

Value

a ‘vector’ of class "mpfr", when one of the first three arguments was.

Author(s)

Martin Maechler

See Also

The documentation of the base function seq; mpfr

Examples

seqMpfr(0, 1, by = mpfr(0.25, prec=88))

seqMpfr(7, 3) # -> default prec.

[Package Rmpfr version 0.8-2 Index]